Pushing styles back to individual classes
